+++ shared/module_loading/pyproject.toml
@@ -24,6 +24,8 @@ classifiers = [
 ]
 
 dependencies = [
+    'importlib_metadata>=6.5;python_version<"3.12"',
+    'importlib_metadata>=7.0;python_version>="3.12"',
 ]
 
 [dependency-groups]
