+++ shared/module_loading/src/airflow_shared/module_loading/__init__.py
@@ -17,10 +17,23 @@
 # under the License.
 from __future__ import annotations
 
+import functools
+import logging
 import pkgutil
-from collections.abc import Callable
+import sys
+from collections import defaultdict
+from collections.abc import Callable, Iterator
 from importlib import import_module
 from typing import TYPE_CHECKING
+
+if sys.version_info >= (3, 12):
+    from importlib import metadata
+else:
+    import importlib_metadata as metadata
+
+log = logging.getLogger(__name__)
+
+EPnD = tuple[metadata.EntryPoint, metadata.Distribution]
 
 if TYPE_CHECKING:
     from types import ModuleType
@@ -89,3 +102,32 @@ def is_valid_dotpath(path: str) -> bool:
     pattern = r"^[a-zA-Z_][a-zA-Z0-9_]*(\.[a-zA-Z_][a-zA-Z0-9_]*)*$"
 
     return bool(re.match(pattern, path))
+
+
+@functools.cache
+def _get_grouped_entry_points() -> dict[str, list[EPnD]]:
+    mapping: dict[str, list[EPnD]] = defaultdict(list)
+    for dist in metadata.distributions():
+        try:
+            for e in dist.entry_points:
+                mapping[e.group].append((e, dist))
+        except Exception as e:
+            log.warning("Error when retrieving package metadata (skipping it): %s, %s", dist, e)
+    return mapping
+
+
+def entry_points_with_dist(group: str) -> Iterator[EPnD]:
+    """
+    Retrieve entry points of the given group.
+
+    This is like the ``entry_points()`` function from ``importlib.metadata``,
+    except it also returns the distribution the entry point was loaded from.
+
+    Note that this may return multiple distributions to the same package if they
+    are loaded from different ``sys.path`` entries. The caller site should
+    implement appropriate deduplication logic if needed.
+
+    :param group: Filter results to only this entrypoint group
+    :return: Generator of (EntryPoint, Distribution) objects for the specified groups
+    """
+    return iter(_get_grouped_entry_points()[group])
