+++ task-sdk/pyproject.toml
@@ -77,6 +77,10 @@ dependencies = [
     "packaging>=24.0",
     "typing-extensions>=4.14.1",
     # End of shared configuration dependencies
+    # Start of shared module-loading dependencies
+    'importlib_metadata>=6.5;python_version<"3.12"',
+    'importlib_metadata>=7.0;python_version>="3.12"',
+    # End of shared module-loading dependencies
 ]
 
 [project.optional-dependencies]
